Move support helper fixture_flows_helper methods out of test/helper
test/helper should be used to store tests of app/helper modules
There is an existing test/support/fixture_methods helper which is
loaded via test_helper.
So methods moved to test/support/fixture_methods and calls to
fixture_flows_helper removed
