+++ test/integration/engine/engine_test_helper.rb
@@ -1,9 +1,6 @@
 require_relative "../../integration_test_helper"
-require_relative "../../helpers/fixture_flows_helper"
 
 class EngineIntegrationTest < ActionDispatch::IntegrationTest
-  include FixtureFlowsHelper
-
   setup do
     setup_fixture_flows
   end
