+++ test/unit/flow_registration_presenter_test.rb
@@ -1,11 +1,8 @@
 require_relative "../test_helper"
-require_relative "../helpers/fixture_flows_helper"
 
 require File.expand_path("../fixtures/smart_answer_flows/flow-sample", __dir__)
 
 class FlowRegistrationPresenterTest < ActiveSupport::TestCase
-  include FixtureFlowsHelper
-
   def setup
     setup_fixture_flows
     @flow = SmartAnswer::FlowSampleFlow.build
